fix_log

+++ train.py
@@ -5,7 +5,7 @@ import logging
 from path import Path
 from utils import custom_transform
 from dataset.KITTI_dataset import KITTI
-from model import DeepVIO, Encoder_CAM
+from model import DeepVIO, Encoder_CAM, FlowVO
 from collections import defaultdict
 from utils.kitti_eval import KITTI_tester
 import numpy as np
@@ -18,7 +18,7 @@ parser.add_argument('--gpu_ids', type=str, default='1', help='gpu ids: e.g. 0  0
 parser.add_argument('--save_dir', type=str, default='./results', help='path to save the result')
 
 parser.add_argument('--train_seq', type=list, default=['00', '01', '02', '04', '06', '08', '09'], help='sequences for training')
-parser.add_argument('--val_seq', type=list, default=['05', '07', '10'], help='sequences for validation')
+parser.add_argument('--val_seq', type=list, default=['05', '07', '10', '01', '02', '04'], help='sequences for validation')
 parser.add_argument('--seed', type=int, default=0, help='random seed')
 
 parser.add_argument('--img_w', type=int, default=512, help='image width')
@@ -49,8 +49,8 @@ parser.add_argument('--Lambda', type=float, default=3e-5, help='penalty factor f
 parser.add_argument('--experiment_name', type=str, default='debug', help='experiment name')
 parser.add_argument('--optimizer', type=str, default='Adam', help='type of optimizer [Adam, SGD]')
 
-# parser.add_argument('--pretrain_flownet',type=str, default='./model_zoo/flownets_bn_EPE2.459.pth.tar', help='wehther to use the pre-trained flownet')
-parser.add_argument('--pretrain_flownet',type=str, default=None, help='wehther to use the pre-trained flownet')
+parser.add_argument('--pretrain_flownet',type=str, default='./model_zoo/flownets_bn_EPE2.459.pth.tar', help='wehther to use the pre-trained flownet')
+# parser.add_argument('--pretrain_flownet',type=str, default=None, help='wehther to use the pre-trained flownet')
 parser.add_argument('--pretrain', type=str, default=None, help='path to the pretrained model')
 parser.add_argument('--hflip', default=False, action='store_true', help='whether to use horizonal flipping as augmentation')
 parser.add_argument('--color', default=False, action='store_true', help='whether to use color augmentations')
@@ -59,6 +59,12 @@ parser.add_argument('--print_frequency', type=int, default=10, help='print frequ
 parser.add_argument('--weighted', default=False, action='store_true', help='whether to use weighted sum')
 parser.add_argument('--patch_size', type=int, default=16, help='patch token size')
 parser.add_argument('--T', type=int, default=2, help='time transformer T=2')
+parser.add_argument('--iter_warmup', type=int, default=10000, help='for iter not epoch')
+parser.add_argument('--iter_joint', type=int, default=20000, help='for iter not epoch')
+parser.add_argument('--iter_fine', type=int, default=20000, help='for iter not epoch')
+parser.add_argument('--epoch_or_iter', action='store_true', default=False, help='true is epoch, false is iter')
+parser.add_argument('--model_type', type=str, default='flowVO', help='path to the pretrained model')
+
 
 args = parser.parse_args()
 
@@ -87,6 +93,11 @@ if args.experiment_name != 'debug':
         "v_f_len": args.v_f_len,
         "i_f_len":args.i_f_len,
         "T": args.T,
+        "epoch_or_iter": args.epoch_or_iter,
+        "iter_warmup": args.iter_warmup,
+        "iter_joint": args.iter_joint,
+        "iter_fine": args.iter_fine,
+        "model_type": args.model_type,
         }
     )
 
@@ -95,25 +106,37 @@ torch.manual_seed(args.seed)
 np.random.seed(args.seed)
 
 def update_status(ep, args, model):
-    if ep < args.epochs_warmup:  # Warmup stage
-        lr = args.lr_warmup
-        selection = 'random'
-        temp = args.temp_init
-        # for param in model.module.Policy_net.parameters(): # Disable the policy network
-        #     param.requires_grad = False
-    elif ep >= args.epochs_warmup and ep < args.epochs_warmup + args.epochs_joint: # Joint training stage
-        lr = args.lr_joint
+    if args.epoch_or_iter:
+        if ep < args.epochs_warmup:  # Warmup stage
+            lr = args.lr_warmup
+            selection = 'random'
+            temp = args.temp_init
+            # for param in model.module.Policy_net.parameters(): # Disable the policy network
+            #     param.requires_grad = False
+        elif ep >= args.epochs_warmup and ep < args.epochs_warmup + args.epochs_joint: # Joint training stage
+            lr = args.lr_joint
+            selection = 'gumbel-softmax'
+            temp = args.temp_init * math.exp(-args.eta * (ep-args.epochs_warmup))
+            # for param in model.module.Policy_net.parameters(): # Enable the policy network
+            #     param.requires_grad = True
+        elif ep >= args.epochs_warmup + args.epochs_joint: # Finetuning stage
+            lr = args.lr_fine
+            selection = 'gumbel-softmax'
+            temp = args.temp_init * math.exp(-args.eta * (ep-args.epochs_warmup))
+    # var ep==iter
+    else:
+        if ep < args.iter_warmup:
+            lr = args.lr_warmup
+        elif ep >= args.iter_warmup and ep < args.iter_warmup + args.iter_joint:
+            lr = args.lr_joint
+        elif ep >= args.iter_warmup + args.iter_joint:
+            lr = args.lr_fine
         selection = 'gumbel-softmax'
         temp = args.temp_init * math.exp(-args.eta * (ep-args.epochs_warmup))
-        # for param in model.module.Policy_net.parameters(): # Enable the policy network
-        #     param.requires_grad = True
-    elif ep >= args.epochs_warmup + args.epochs_joint: # Finetuning stage
-        lr = args.lr_fine
-        selection = 'gumbel-softmax'
-        temp = args.temp_init * math.exp(-args.eta * (ep-args.epochs_warmup))
+
     return lr, selection, temp
 
-def train(model, optimizer, train_loader, selection, temp, logger, ep, p=0.5, weighted=False):
+def train(model, optimizer, train_loader, selection, temp, logger, ep, iter, p=0.5, weighted=False):
     
     mse_losses = []
     penalties = []
@@ -128,7 +151,7 @@ def train(model, optimizer, train_loader, selection, temp, logger, ep, p=0.5, we
 
         optimizer.zero_grad()
         # 模型结果
-        poses, attn_map, decisions, probs = model(imgs)
+        poses, decisions, probs = model(imgs)
 
         
         if not weighted:
@@ -146,22 +169,28 @@ def train(model, optimizer, train_loader, selection, temp, logger, ep, p=0.5, we
         
         loss.backward()
         optimizer.step()
+
+        iter = iter + 1
         
         if i % args.print_frequency == 0: 
-            message = f'Epoch: {ep}, iters: {i}/{data_len}, pose loss: {pose_loss.item():.6f}, penalty: {penalty.item():.6f}, loss: {loss.item():.6f}'
+            message = f'Epoch: {ep}, iter: {iter}/{data_len}, pose loss: {pose_loss.item():.6f}, penalty: {penalty.item():.6f}, loss: {loss.item():.6f}'
             print(message)
             if args.experiment_name != 'debug':
-                wandb.log({"Epoch": ep, "iters": i, "pose loss": pose_loss.item(),"penalty": penalty, "angle loss": angle_loss.item(), "translation loss": translation_loss.item(), "loss": loss.item()})
+                wandb.log({"Epoch": ep, "iter": iter, "pose loss": pose_loss.item(),"penalty": penalty, "angle loss": angle_loss.item(), "translation loss": translation_loss.item(), "loss": loss.item()})
             logger.info(message)
 
         mse_losses.append(pose_loss.item())
         penalties.append(penalty.item())
+        if iter > (args.iter_warmup + args.iter_warmup + args.iter_joint):
+            break
 
-    return np.mean(mse_losses), np.mean(penalties)
+    return np.mean(mse_losses), np.mean(penalties), iter
+
 
 
 def main():
-
+    iter = 0
+    ep = 0 
     # Create Dir
     experiment_dir = Path('./results')
     experiment_dir.mkdir_p()
@@ -222,8 +251,12 @@ def main():
     tester = KITTI_tester(args)
 
     # Model initialization
-    # model = DeepVIO(args)
-    model = Encoder_CAM(args)
+    if args.model_type=='flowVO':
+        model = FlowVO(args)
+    elif args.model_type=='ts_cam':
+        model = Encoder_CAM(args)
+    elif args.model_type=='deepvio':
+        model = DeepVIO(args)
 
     # Continual training or not
     if args.pretrain is not None:
@@ -235,7 +268,7 @@ def main():
         logger.info('Training from scratch')
     
     # Use the pre-trained flownet or not
-    if args.pretrain_flownet and args.pretrain is None:
+    if args.model_type=='flowVO' and args.pretrain_flownet and args.pretrain is None:
         pretrained_w = torch.load(args.pretrain_flownet, map_location='cpu')
         model_dict = model.Feature_net.state_dict()
         update_dict = {k: v for k, v in pretrained_w['state_dict'].items() if k in model_dict}
@@ -246,8 +279,9 @@ def main():
     model.cuda(gpu_ids[0])
     model = torch.nn.DataParallel(model, device_ids = gpu_ids)
 
-    pretrain = args.pretrain 
-    init_epoch = int(pretrain[-7:-4])+1 if args.pretrain is not None else 0    
+    pretrain = args.pretrain
+    init_epoch = int(pretrain[-7:-4])+1 if args.pretrain is not None else 0
+    iter = init_epoch * (len(train_loader) // args.batch_size + 1) if args.pretrain is not None else 0
     
     # Initialize the optimizer
     if args.optimizer == 'SGD':
@@ -258,31 +292,33 @@ def main():
     
     best = 10000
 
-    for ep in range(init_epoch, args.epochs_warmup+args.epochs_joint+args.epochs_fine):
+    # 如果还是想使用ep来控制，那么就将main函数上面的ep=0给注释掉 最底下的ep=ep+1也给注释掉
+    # for ep in range(init_epoch, args.epochs_warmup+args.epochs_joint+args.epochs_fine):
+    while(iter < args.iter_warmup+args.iter_joint+args.iter_fine):
         
-        lr, selection, temp = update_status(ep, args, model)
+        lr, selection, temp = update_status(iter, args, model)
         optimizer.param_groups[0]['lr'] = lr
-        message = f'Epoch: {ep}, lr: {lr}, selection: {selection}, temperaure: {temp:.5f}'
+        message = f'Epoch: {ep}, iter: {iter} lr: {lr}, selection: {selection}, temperaure: {temp:.5f}'
         print(message)
         logger.info(message)
 
         model.train()
-        avg_pose_loss, avg_penalty_loss = train(model, optimizer, train_loader, selection, temp, logger, ep, p=0.5)
+        avg_pose_loss, avg_penalty_loss, iter = train(model, optimizer, train_loader, selection, temp, logger, ep, iter=iter, p=0.5)
         
         # Save the model after training
-        if(os.path.isfile(f'{checkpoints_dir}/{(ep-1):003}.pth')):
+        if(os.path.isfile(f'{checkpoints_dir}/{(ep-1):003}.pth') and ep%10!=0):
             os.remove(f'{checkpoints_dir}/{(ep-1):003}.pth')
         torch.save(model.module.state_dict(), f'{checkpoints_dir}/{ep:003}.pth')
         # Save the model after training
-        message = f'Epoch {ep} training finished, pose loss: {avg_pose_loss:.6f}, penalty_loss: {avg_penalty_loss:.6f}, model saved'
+        message = f'Epoch {ep} iter: {iter} training finished, pose loss: {avg_pose_loss:.6f}, penalty_loss: {avg_penalty_loss:.6f}, model saved'
         print(message)
         logger.info(message)
         
-        if ep > args.epochs_warmup+args.epochs_joint or (ep > 10 and ep % 2 == 0):
+        if ep % 5 == 0 or iter == args.iter_warmup+args.iter_joint+args.iter_fine-1:
             # Evaluate the model
             print('Evaluating the model')
             logger.info('Evaluating the model')
-            with torch.no_grad(): 
+            with torch.no_grad():  
                 model.eval()
                 errors = tester.eval(model, selection='gumbel-softmax', num_gpu=len(gpu_ids))
         
@@ -297,25 +333,46 @@ def main():
                 if best < 10:
                     torch.save(model.module.state_dict(), f'{checkpoints_dir}/best_{best:.2f}.pth')
 
-            message = "Epoch {} evaluation Seq. 05 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, round(errors[0]['t_rel'], 4), round(errors[0]['r_rel'], 4), round(errors[0]['t_rmse'], 4), round(errors[0]['r_rmse'], 4))
+            message = "Epoch {} iter: {} evaluation Seq. 05 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, iter, round(errors[0]['t_rel'], 4), round(errors[0]['r_rel'], 4), round(errors[0]['t_rmse'], 4), round(errors[0]['r_rmse'], 4))
             logger.info(message)
             print(message)
             if args.experiment_name != 'debug':
-                wandb.log({"5. t_rel": round(errors[0]['t_rel'], 4), "5. r_rel": round(errors[0]['r_rel'], 4), "5. t_rmse": round(errors[0]['t_rmse'], 4), "5. r_rmse": round(errors[0]['r_rmse'], 4)})
+                wandb.log({"Epoch": ep, "iter":iter, "5. t_rel": round(errors[0]['t_rel'], 4), "5. r_rel": round(errors[0]['r_rel'], 4), "5. t_rmse": round(errors[0]['t_rmse'], 4), "5. r_rmse": round(errors[0]['r_rmse'], 4)})
 
-            message = "Epoch {} evaluation Seq. 07 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, round(errors[1]['t_rel'], 4), round(errors[1]['r_rel'], 4), round(errors[1]['t_rmse'], 4), round(errors[1]['r_rmse'], 4))
+            message = "Epoch {} iter: {} evaluation Seq. 07 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, iter, round(errors[1]['t_rel'], 4), round(errors[1]['r_rel'], 4), round(errors[1]['t_rmse'], 4), round(errors[1]['r_rmse'], 4))
             logger.info(message)
             print(message)
             if args.experiment_name != 'debug':
-                wandb.log({"7. t_rel": round(errors[1]['t_rel'], 4), "7. r_rel": round(errors[1]['r_rel'], 4), "7. t_rmse": round(errors[1]['t_rmse'], 4), "7. r_rmse": round(errors[1]['r_rmse'], 4)})
+                wandb.log({"Epoch": ep, "iter":iter, "7. t_rel": round(errors[1]['t_rel'], 4), "7. r_rel": round(errors[1]['r_rel'], 4), "7. t_rmse": round(errors[1]['t_rmse'], 4), "7. r_rmse": round(errors[1]['r_rmse'], 4)})
 
-            message = "Epoch {} evaluation Seq. 10 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, round(errors[2]['t_rel'], 4), round(errors[2]['r_rel'], 4), round(errors[2]['t_rmse'], 4), round(errors[2]['r_rmse'], 4))
+            message = "Epoch {} iter: {} evaluation Seq. 10 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, iter, round(errors[2]['t_rel'], 4), round(errors[2]['r_rel'], 4), round(errors[2]['t_rmse'], 4), round(errors[2]['r_rmse'], 4))
             logger.info(message)
             print(message)
             if args.experiment_name != 'debug':
-                wandb.log({"10. t_rel": round(errors[2]['t_rel'], 4), "10. r_rel": round(errors[2]['r_rel'], 4), "10. t_rmse": round(errors[2]['t_rmse'], 4), "10. r_rmse": round(errors[2]['r_rmse'], 4)})
+                wandb.log({"Epoch": ep, "iter":iter, "10. t_rel": round(errors[2]['t_rel'], 4), "10. r_rel": round(errors[2]['r_rel'], 4), "10. t_rmse": round(errors[2]['t_rmse'], 4), "10. r_rmse": round(errors[2]['r_rmse'], 4)})
+            
+            message = "Epoch {} iter: {} evaluation Seq. 01 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, iter, round(errors[3]['t_rel'], 4), round(errors[3]['r_rel'], 4), round(errors[3]['t_rmse'], 4), round(errors[3]['r_rmse'], 4))
             logger.info(message)
             print(message)
+            if args.experiment_name != 'debug':
+                wandb.log({"Epoch": ep, "iter":iter, "1. t_rel": round(errors[3]['t_rel'], 4), "01. r_rel": round(errors[3]['r_rel'], 4), "01. t_rmse": round(errors[3]['t_rmse'], 4), "01. r_rmse": round(errors[3]['r_rmse'], 4)})
+            
+            message = "Epoch {} iter: {} evaluation Seq. 02 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, iter, round(errors[4]['t_rel'], 4), round(errors[4]['r_rel'], 4), round(errors[4]['t_rmse'], 4), round(errors[4]['r_rmse'], 4))
+            logger.info(message)
+            print(message)
+            if args.experiment_name != 'debug':
+                wandb.log({"Epoch": ep, "iter":iter, "2. t_rel": round(errors[4]['t_rel'], 4), "02. r_rel": round(errors[4]['r_rel'], 4), "02. t_rmse": round(errors[4]['t_rmse'], 4), "02. r_rmse": round(errors[4]['r_rmse'], 4)})
+            
+            message = "Epoch {} iter: {} evaluation Seq. 04 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, iter, round(errors[5]['t_rel'], 4), round(errors[5]['r_rel'], 4), round(errors[5]['t_rmse'], 4), round(errors[5]['r_rmse'], 4))
+            logger.info(message)
+            print(message)
+            if args.experiment_name != 'debug':
+                wandb.log({"Epoch": ep, "iter":iter, "4. t_rel": round(errors[5]['t_rel'], 4), "04. r_rel": round(errors[5]['r_rel'], 4), "04. t_rmse": round(errors[5]['t_rmse'], 4), "04. r_rmse": round(errors[5]['r_rmse'], 4)})
+            
+            logger.info(message)
+            print(message)
+            
+        ep += 1
     
     message = f'Training finished, best t_rel: {best:.4f}'
     logger.info(message)
@@ -323,7 +380,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
-
-
-
